Switch annotation to class pattern
This will make it easier to use a registry pattern in the future (i.e., to auto-group things, like what's being done manually in the MetricSets)
