+++ rexmex/utils.py
@@ -2,6 +2,12 @@ from functools import wraps
 
 import numpy as np
 
+__all__ = [
+    "binarize",
+    "normalize",
+    "Annotator",
+]
+
 
 def binarize(metric):
     """
@@ -47,3 +53,19 @@ def normalize(metric):
         return score
 
     return metric_wrapper
+
+
+class Annotator:
+    """A class to wrap annotations to make the registry pattern easier later."""
+
+    def annotate(self, *, name: str, lower: float, upper: float, higher_is_better: bool):
+        """Annotate a classification function."""
+
+        def _wrapper(func):
+            func.name = name
+            func.lower = lower
+            func.upper = upper
+            func.higher_is_better = higher_is_better
+            return func
+
+        return _wrapper
